LayoutTests/http/tests/websocket/tests/hybi/error-event-ready-state.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../../js-test-resources/js-test.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Test that readyState is CLOSED within onerror event");

window.jsTestIsAsync = true;

var ws = new WebSocket("ws://localhost:8880/non-existent-url");

ws.onerror = function(event) {
    debug("onerror() was called.");
    shouldBe("ws.readyState", "3");
    finishJSTest();
};

</script>
<script src="../../../../js-test-resources/js-test.js"></script>
</body>
</html>
